Another refactor. Getting more confident here, but this is taking MUCH longer than we planned
