+++ DevOpsMetrics/DevOpsMetrics.Service/DataAccess/PullRequestDA.cs
@@ -1,5 +1,6 @@
 ﻿using DevOpsMetrics.Service.Models;
 using DevOpsMetrics.Service.Models.AzureDevOps;
+using DevOpsMetrics.Service.Models.Common;
 using DevOpsMetrics.Service.Models.GitHub;
 using Newtonsoft.Json;
 using System.Collections.Generic;
