+++ DevOpsMetrics/DevOpsMetrics.Service/Models/Common/Committer.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace DevOpsMetrics.Service.Models
+namespace DevOpsMetrics.Service.Models.Common
 {
     public class Committer
     {
