+++ DevOpsMetrics/DevOpsMetrics.Web/Views/Shared/_AzureDevOpsLeadTimeForChanges.cshtml
@@ -1,4 +1,4 @@
-﻿@using DevOpsMetrics.Service.Models 
+﻿@using DevOpsMetrics.Service.Models.Common 
 @model LeadTimeForChangesPartialViewModel
 
 <div class="chartContainer vertical flat">
